refactor: remove v1 directory structure and simplify API endpoints
- Remove unnecessary v1 directory hierarchy from backend API structure
- Move market_data.py and data_collection.py directly under app/api/
- Update API endpoint prefixes from /api/v1 to /api
- Update frontend API calls to use new endpoint structure
- Update test files to reference new API paths
- Maintain all existing functionality while simplifying codebase

Benefits:
- Cleaner, more intuitive API structure
- Reduced unnecessary nesting
- Improved maintainability
- Consistent with YAGNI principle (You Aren't Gonna Need It)

Files changed:
Backend:
- backend/app/main.py: Updated imports and routing prefixes
- backend/app/api/market_data.py: Moved from v1 subdirectory
- backend/app/api/data_collection.py: Moved from v1 subdirectory
- backend/tests/unit/test_data_collection_api.py: Updated test URLs

Frontend:
- frontend/app/api/data/candlesticks/route.ts: Updated backend API URL
- frontend/app/data/page.tsx: Updated API endpoint calls
