--- backend/app/api/v1/__init__.py
@@ -1,3 +1,0 @@
-"""
-API v1 パッケージ
-"""
